Back

+++ trabalho_IA/src/br/com/academic/genetic/model/Product.java
@@ -1,15 +1,13 @@
 package br.com.academic.genetic.model;
 
-import java.math.BigDecimal;
-
 public class Product {
 
 	private final String name;
 	private final Double volume;
-	private final BigDecimal price;
+	private final Double price;
 	private final String imageName;
 
-	public Product(String name, Double volume, BigDecimal price, String imageName) {
+	public Product(String name, Double volume, Double price, String imageName) {
 		this.name = name;
 		this.volume = volume;
 		this.price = price;
@@ -24,7 +22,7 @@ public class Product {
 		return volume;
 	}
 
-	public BigDecimal getPrice() {
+	public Double getPrice() {
 		return price;
 	}
